+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Tqdm workaround (https://github.com/tqdm/tqdm/issues/777).
 - Ledger Live parser: unrecognised operation type 'IN'.
 - TradeSatoshi parser: calculations rounded to 8 decimal places.
+- Trezor parser: "self" payment exception.
 ### Added
 - Conversion tool: added parser for CoinTracking.info accounting data.
 - Conversion tool: added parser for Gravity (Bitstocks) exchange.